Frontend custom enchant builder bijna af. Alleen nog YAML naar JSON loaden toevoegen
