+++ frontend/package.json
@@ -19,7 +19,8 @@
     "react-router-dom": "^7.0.1",
     "react-scripts": "^5.0.1",
     "three": "^0.170.0",
-    "web-vitals": "^4.2.4"
+    "web-vitals": "^4.2.4",
+    "yaml": "^2.6.1"
   },
   "scripts": {
     "start": "react-scripts start",
