+++ frontend/src/components/custom_enchants/content_pages/custom_enchant_builder/CustomEnchantBuilderContent.js
@@ -1,17 +1,26 @@
-import React, {useState} from "react";
+import React, {useState, useEffect} from "react";
 import InputField from "../InputField";
 import SelectField from "../SelectField";
 import "../../../../styles/custom_enchants/CustomEnchants.css";
-import { versions, enchantment_targets, enchantment_tags, enchantments, triggers, command_parameters } from "../../../../data";
+import { versions } from "../../../../data/versions";
+import { enchantment_targets } from "../../../../data/targets"
+import { enchantment_tags } from "../../../../data/tags"
+import { enchantments } from "../../../../data/enchantments";
+import { triggers } from "../../../../data/triggers";
+import { command_parameters } from "../../../../data/commandParameters";
 import AddableSelectField from "../AddableSelectField";
 import CheckboxField from "../CheckboxField";
 import TriggerSelectField from "./TriggerSelectField";
 import LevelCreationField from "./LevelCreationField"
-
+import { checkConstraints } from "../../../../util/constraints";
+import { jsonToYaml } from "../../../../util/yamlParser";
 
 const defaultFormState = {
   minecraft_version: versions[0],
   enchantment_name: "",
+  targets: [],
+  tags: [],
+  conflicts_with: [],
   anvil_cost: 2,
   in_enchanting_table: true,
   weight: 10,
@@ -19,6 +28,7 @@ const defaultFormState = {
   min_cost_incr: 1,
   max_cost_base: 5,
   max_cost_incr: 1,
+  cooldown_message: "&7You, &6%player%&7, have to wait %time_left% or %time_left_full_out% before you can use %enchantment% again!",
   levels: [
     {
       cooldown: 60,
@@ -30,54 +40,26 @@ const defaultFormState = {
   triggers: []
 }
 
-const restrictions = {
-  weight: {
-    min: 1, 
-    max: 1024, 
-    parse: (n) => parseInt(n) 
-  },
-  min_cost_base: {
-    min: 1, 
-    max: 30, 
-    parse: (n) => parseInt(n) 
-  },
-  min_cost_incr: {
-    min: 0,
-    max: 30,
-    parse: (n) => parseInt(n)
-  },
-  max_cost_base: {
-    min: 1, 
-    max: 30, 
-    parse: (n) => parseInt(n) 
-  },
-  max_cost_incr: {
-    min: 0,
-    max: 30,
-    parse: (n) => parseInt(n)
-  }
-}
-
 
 const CustomEnchantBuilderContent = () => {
-  const [formState, setFormState] = useState(defaultFormState);
+  const [formState, setFormState] = useState(() => {
+    const storedData = localStorage.getItem('formState');
+    return storedData ? JSON.parse(storedData) : defaultFormState
+  });
+  const [errors, setErrors] = useState([]);
+  const [copySucces, setCopySuccess] = useState(false);
+
+  useEffect(() => {
+    localStorage.setItem('formState', JSON.stringify(formState));
+}, [formState]);
+
 
   const handleChange = (event) => {    
     const { name, value } = event.target;
 
-    const restriction = restrictions[name];
-    let parsedValue = undefined;
-    if (restriction && value.trim() !== "") {
-      parsedValue = restriction.parse(value);
-      if (parsedValue < restriction.min || parsedValue > restriction.max)
-        return;
-    } else {
-      parsedValue = value;
-    }
-    
     setFormState((prevState) => ({
       ...prevState,
-      [name]: parsedValue,
+      [name]: value,
     }));
   };
 
@@ -110,6 +92,16 @@ const CustomEnchantBuilderContent = () => {
     }
   }
 
+  const getYamlOutput = () => {
+    const errors = checkConstraints(formState);
+    setErrors(errors);
+    if (errors.length > 0) return;
+
+    navigator.clipboard.writeText(jsonToYaml(formState));
+    setCopySuccess(true);
+    setTimeout(() => setCopySuccess(false), 3000);
+  } 
+
   
 
   return (
@@ -119,9 +111,8 @@ const CustomEnchantBuilderContent = () => {
       </p>
       <div className="content-box">
         <h2 className="content-box-title">General Information</h2>
-        <div className="field-container">
-          <SelectField 
-            label="Minecraft Version: " 
+        <SelectField 
+            label="Minecraft Version" 
             description= "The Minecraft version you will use this enchantment on"
             options={versions.map((item) => ({ value: item, label: item }))} 
             name="minecraft_version"
@@ -129,36 +120,38 @@ const CustomEnchantBuilderContent = () => {
             onChange={handleChange}
           />
           <InputField 
-            label="Enchantment Name: " 
+            label="Enchantment Name" 
             description="The name you want the enchantment to have"
             placeholder=""
             name="enchantment_name"
             value={formState.enchantment_name}
             onChange={handleChange}
           />
-        </div>
       </div>
       <div className="content-box">
         <h2 className="content-box-title">Enchantment Definition</h2>
         <AddableSelectField
           name = "targets"
-          label = "Targets: "
+          label = "Targets"
           description= "The target items for your enchantment"
           options={enchantment_targets}
+          values={formState.targets}
           onChange={handleAddableSelectboxChange}
         />
         <AddableSelectField
           name="tags"
-          label = "Tags: "
+          label = "Tags"
           description= "Modifiers to customize your enchantment's vanilla behaviour"
           options={enchantment_tags}
+          values={formState.tags}
           onChange={handleAddableSelectboxChange}
         />
         <AddableSelectField
           name="conflicts_with"
-          label = "Conflicts with: "
+          label = "Conflicts with"
           description= "Enchantments that cannot be combined with your enchantment on the same item."
           options={enchantments}
+          values={formState.conflicts_with}
           customOptionsAllowed={true}
           onChange={handleAddableSelectboxChange}
         />
@@ -254,6 +247,14 @@ const CustomEnchantBuilderContent = () => {
 
         <div className="content-box">
           <h2 className="content-box-title">Levels</h2>
+          <InputField
+            label="Cooldown Message"
+            description="The message to be sent to the player when the enchantment is on cooldown from triggering again (leave empty if you dont want any message to be shown)"
+            placeholder=""
+            name="cooldown_message"
+            value={formState.cooldown_message}
+            onChange={handleChange}
+          />
           <LevelCreationField
             levels={formState.levels}
             parameters={filteredParameters()}
@@ -262,18 +263,30 @@ const CustomEnchantBuilderContent = () => {
         </div>
         <br/>
         <br/>
-        <br/>
-        <br/>
         <div className="content-box">
           <h2 className="content-box-title">Output</h2>
-          <button className="add-btn-text" onClick={""}>
-            Get Yaml Output
-          </button>
+          <button
+                className={`add-btn-text ${copySucces ? 'copy-success' : ''}`}
+                onClick={getYamlOutput}
+                disabled={copySucces}
+            >
+                {copySucces ? "Output copied to clipboard!": "Get Yaml Output"}
+            </button>
           <button className="add-btn-text red" onClick={clearAllInput}>
             Clear All Input
           </button>
-        </div>
 
+          {errors.length > 0 && (
+            <div className="error-list">
+              <h3 className="error-list-title">Please fix the following errors:</h3>
+              <ul>
+                {errors.map((error, index) => (
+                  <li key={index} className="error-item">{error}</li>
+                ))}
+              </ul>
+            </div>
+          )}
+        </div>
       </div>
     </div>
   );
