+++ frontend/src/data/commandFunctions.js
@@ -1,0 +1,14 @@
+export const command_functions = [
+    {
+      name: "MUL(x1, x2, ...)",
+      description: "Multiply the given values together"
+    },
+    {
+      name: "ADD(x1, x2, ...)",
+      description: "Adds the given values together"
+    },
+    {
+      name: "SUB(x1, x2, ...)",
+      description: "Subtracts the given values from each other"
+    }
+  ]
