+++ frontend/src/data/damageCauses.js
@@ -1,0 +1,34 @@
+export const damage_causes = [
+    { name: "kill", label: "Kill" },
+    { name: "world_border", label: "World Border" },
+    { name: "contact", label: "Contact" },
+    { name: "entity_attack", label: "Entity Attack" },
+    { name: "entity_sweep_attack", label: "Entity Sweep Attack" },
+    { name: "projectile", label: "Projectile" },
+    { name: "suffocation", label: "Suffocation" },
+    { name: "fall", label: "Fall" },
+    { name: "fire", label: "Fire" },
+    { name: "fire_tick", label: "Fire Tick" },
+    { name: "melting", label: "Melting" },
+    { name: "lava", label: "Lava" },
+    { name: "drowning", label: "Drowning" },
+    { name: "block_explosion", label: "Block Explosion" },
+    { name: "entity_explosion", label: "Entity Explosion" },
+    { name: "void", label: "Void" },
+    { name: "lightning", label: "Lightning" },
+    { name: "suicide", label: "Suicide" },
+    { name: "starvation", label: "Starvation" },
+    { name: "poison", label: "Poison" },
+    { name: "magic", label: "Magic" },
+    { name: "wither", label: "Wither" },
+    { name: "falling_block", label: "Falling Block" },
+    { name: "thorns", label: "Thorns" },
+    { name: "dragon_breath", label: "Dragon Breath" },
+    { name: "custom", label: "Custom" },
+    { name: "fly_into_wall", label: "Fly Into Wall" },
+    { name: "hot_floor", label: "Hot Floor" },
+    { name: "cramming", label: "Cramming" },
+    { name: "dryout", label: "Dryout" },
+    { name: "freeze", label: "Freeze" },
+    { name: "sonic_boom", label: "Sonic Boom" },
+  ];
