+++ frontend/src/data/enchantments.js
@@ -1,0 +1,212 @@
+export const enchantments = [
+    {
+      "name": "aqua_affinity",
+      "label": "Aqua Affinity",
+      "description": "Increases underwater mining speed"
+    },
+    {
+      "name": "bane_of_arthropods",
+      "label": "Bane of Arthropods",
+      "description": "Increases damage against arthropod enemies (e.g., spiders, silverfish)"
+    },
+    {
+      "name": "blast_protection",
+      "label": "Blast Protection",
+      "description": "Reduces damage from explosions"
+    },
+    {
+      "name": "channeling",
+      "label": "Channeling",
+      "description": "Causes your Trident to summon a lightning bolt when thrown during a thunderstorm"
+    },
+    {
+      "name": "curse_of_binding",
+      "label": "Curse of Binding",
+      "description": "Prevents armor from being removed once equipped"
+    },
+    {
+      "name": "curse_of_vanishing",
+      "label": "Curse of Vanishing",
+      "description": "Causes items to disappear when you die"
+    },
+    {
+      "name": "depth_strider",
+      "label": "Depth Strider",
+      "description": "Increases movement speed underwater"
+    },
+    {
+      "name": "efficiency",
+      "label": "Efficiency",
+      "description": "Increases mining speed"
+    },
+    {
+      "name": "feather_falling",
+      "label": "Feather Falling",
+      "description": "Reduces fall damage"
+    },
+    {
+      "name": "fire_aspect",
+      "label": "Fire Aspect",
+      "description": "Sets targets on fire when hit"
+    },
+    {
+      "name": "fire_protection",
+      "label": "Fire Protection",
+      "description": "Reduces damage from fire"
+    },
+    {
+      "name": "flame",
+      "label": "Flame",
+      "description": "Turns arrows into flaming arrows"
+    },
+    {
+      "name": "fortune",
+      "label": "Fortune",
+      "description": "Increases the number of items dropped when mining certain blocks"
+    },
+    {
+      "name": "frost_walker",
+      "label": "Frost Walker",
+      "description": "Creates ice blocks while walking over water and grants immunity to damage from certain blocks"
+    },
+    {
+      "name": "impaling",
+      "label": "Impaling",
+      "description": "Increases attack damage of the Trident against underwater mobs"
+    },
+    {
+      "name": "infinity",
+      "label": "Infinity",
+      "description": "Arrows are not consumed when shooting"
+    },
+    {
+      "name": "knockback",
+      "label": "Knockback",
+      "description": "Increases knockback dealt to enemies"
+    },
+    {
+      "name": "looting",
+      "label": "Looting",
+      "description": "Increases items dropped by mobs"
+    },
+    {
+      "name": "loyalty",
+      "label": "Loyalty",
+      "description": "Returns the Trident after being thrown"
+    },
+    {
+      "name": "luck_of_the_sea",
+      "label": "Luck of the Sea",
+      "description": "Increases chance of catching valuable items while fishing"
+    },
+    {
+      "name": "lure",
+      "label": "Lure",
+      "description": "Decreases time to catch fish"
+    },
+    {
+      "name": "mending",
+      "label": "Mending",
+      "description": "Repairs the item using experience orbs"
+    },
+    {
+      "name": "multishot",
+      "label": "Multishot",
+      "description": "Shoots multiple arrows or projectiles at once"
+    },
+    {
+      "name": "piercing",
+      "label": "Piercing",
+      "description": "Arrows fired can pierce through multiple entities"
+    },
+    {
+      "name": "power",
+      "label": "Power",
+      "description": "Increases arrow damage"
+    },
+    {
+      "name": "projectile_protection",
+      "label": "Projectile Protection",
+      "description": "Reduces damage from projectiles"
+    },
+    {
+      "name": "protection",
+      "label": "Protection",
+      "description": "Reduces overall damage"
+    },
+    {
+      "name": "punch",
+      "label": "Punch",
+      "description": "Increases arrow knockback"
+    },
+    {
+      "name": "quick_charge",
+      "label": "Quick Charge",
+      "description": "Decreases time to reload a crossbow"
+    },
+    {
+      "name": "respiration",
+      "label": "Respiration",
+      "description": "Extends underwater breathing time"
+    },
+    {
+      "name": "riptide",
+      "label": "Riptide",
+      "description": "Propels the player forward when the Trident is thrown in water or rain"
+    },
+    {
+      "name": "sharpness",
+      "label": "Sharpness",
+      "description": "Increases melee attack damage"
+    },
+    {
+      "name": "silk_touch",
+      "label": "Silk Touch",
+      "description": "Mined blocks drop themselves instead of their usual items"
+    },
+    {
+      "name": "smite",
+      "label": "Smite",
+      "description": "Increases damage against undead mobs"
+    },
+    {
+      "name": "soul_speed",
+      "label": "Soul Speed",
+      "description": "Increases movement speed on Soul Sand and Soul Soil"
+    },
+    {
+      "name": "sweeping_edge",
+      "label": "Sweeping Edge",
+      "description": "Increases damage of sweep attacks"
+    },
+    {
+      "name": "swift_sneak",
+      "label": "Swift Sneak",
+      "description": "Increases movement speed while crouching"
+    },
+    {
+      "name": "thorns",
+      "label": "Thorns",
+      "description": "Damages attackers when they deal melee damage"
+    },
+    {
+      "name": "unbreaking",
+      "label": "Unbreaking",
+      "description": "Increases item durability by reducing the chance of durability loss"
+    },
+    {
+      "name": "density",
+      "label": "Density",
+      "description": "Increases damage dealt by mace smash attacks based on fall distance"
+    },
+    {
+      "name": "breach",
+      "label": "Breach",
+      "description": "Reduces the effectiveness of a target's armor when using a mace"
+    },
+    {
+      "name": "wind_burst",
+      "label": "Wind Burst",
+      "description": "Grants a bounce-back effect after a successful mace smash attack, allowing for chained attacks"
+    }
+]
