+++ frontend/src/data/primeCauses.js
@@ -1,0 +1,30 @@
+export const prime_causes = [
+    {
+      name: "fire",
+      label: "Fire",
+    },
+    {
+      name: "redstone",
+      label: "Redstone",
+    },
+    {
+      name: "player",
+      label: "Player",
+    },
+    {
+      name: "explosion",
+      label: "Explosion",
+    },
+    {
+      name: "projectile",
+      label: "Projectile",
+    },
+    {
+      name: "block_break",
+      label: "Block Break",
+    },
+    {
+      name: "dispenser",
+      label: "Dispenser",
+    },
+]
