+++ frontend/src/data/tags.js
@@ -1,0 +1,132 @@
+export const enchantment_tags = [
+    {
+        name: "TRADEABLE",
+        label: "Tradeable",
+        description: "Enchantement can be found trading with villagers"
+    },
+    {
+        name: "DOUBLE_TRADE_PRICE",
+        label: "Double Trade Price",
+        description: "Doubles the price of this enchantment when trading"
+    },
+    {
+        name: "ON_MOB_SPAWN_EQUIPMENT",
+        label: "On Mob Spawn Equipment",
+        description: "Enchantment can be found on mobs' armor pieces"
+    },
+    {
+        name: "ON_TRADED_EQUIPMENT",
+        label: "On Traded Equipment",
+        description: "Appears on equipment that is bought from villagers"
+    },
+    {
+        name: "ON_RANDOM_LOOT",
+        label: "On Random Loot",
+        description: "Can be discovered in loot chests or dropped by mobs"
+    },
+    {
+        name: "CURSE",
+        label: "Curse",
+        description: "Applies the curse effect to the enchantment"
+    },
+    {
+        name: "SMELTS_LOOT",
+        label: "Smelts Loot",
+        description: "Automatically smelts loot dropped from blocks or entities"
+    },
+    {
+        name: "PREVENTS_BEE_SPAWNS_WHEN_MINING",
+        label: "Prevents Bee Spawns When Mining",
+        description: "Stops bees from spawning when breaking hive-related blocks"
+    },
+    {
+        name: "PREVENTS_DECORATED_POT_SHATTERING",
+        label: "Prevents Decorated Pot Shattering",
+        description: "Prevents decorated pots from breaking when mishandled"
+    },
+    {
+        name: "PREVENTS_ICE_MELTING",
+        label: "Prevents Ice Melting",
+        description: "Prevents ice blocks from melting near heat sources"
+    },
+    {
+        name: "PREVENTS_INFESTED_SPAWNS",
+        label: "Prevents Infested Spawns",
+        description: "Stops infested blocks from spawning silverfish or other pests"
+    },
+    {
+        name: "TREASURE",
+        label: "Treasure",
+        description: "Can be found in lost treasure"
+    },
+    {
+        name: "TRADES_DESERT_COMMON",
+        label: "Trades Desert Common",
+        description: "Commonly traded in desert biome villages"
+    },
+    {
+        name: "TRADES_JUNGLE_COMMON",
+        label: "Trades Jungle Common",
+        description: "Commonly traded in jungle biome villages"
+    },
+    {
+        name: "TRADES_PLAINS_COMMON",
+        label: "Trades Plains Common",
+        description: "Commonly traded in plains biome villages"
+    },
+    {
+        name: "TRADES_SAVANNA_COMMON",
+        label: "Trades Savanna Common",
+        description: "Commonly traded in savanna biome villages"
+    },
+    {
+        name: "TRADES_SNOW_COMMON",
+        label: "Trades Snow Common",
+        description: "Commonly traded in snow biome villages"
+    },
+    {
+        name: "TRADES_SWAMP_COMMON",
+        label: "Trades Swamp Common",
+        description: "Commonly traded in swamp biome villages"
+    },
+    {
+        name: "TRADES_TAIGA_COMMON",
+        label: "Trades Taiga Common",
+        description: "Commonly traded in taiga biome villages"
+    },
+    {
+        name: "TRADES_DESERT_SPECIAL",
+        label: "Trades Desert Special",
+        description: "Special item traded in desert biome villages"
+    },
+    {
+        name: "TRADES_JUNGLE_SPECIAL",
+        label: "Trades Jungle Special",
+        description: "Special item traded in jungle biome villages"
+    },
+    {
+        name: "TRADES_PLAINS_SPECIAL",
+        label: "Trades Plains Special",
+        description: "Special item traded in plains biome villages"
+    },
+    {
+        name: "TRADES_SAVANNA_SPECIAL",
+        label: "Trades Savanna Special",
+        description: "Special item traded in savanna biome villages"
+    },
+    {
+        name: "TRADES_SNOW_SPECIAL",
+        label: "Trades Snow Special",
+        description: "Special item traded in snow biome villages"
+    },
+    {
+        name: "TRADES_SWAMP_SPECIAL",
+        label: "Trades Swamp Special",
+        description: "Special item traded in swamp biome villages"
+    },
+    {
+        name: "TRADES_TAIGA_SPECIAL",
+        label: "Trades Taiga Special",
+        description: "Special item traded in taiga biome villages"
+    }
+];
