+++ frontend/src/data/triggerConditionDescriptions.js
@@ -1,0 +1,11 @@
+export const trigger_condition_descriptions = {
+    block: "Select which blocks should trigger the enchantment (All by default)",
+    entity: "Select which entities should trigger the enchantment (All by default)",
+    mob: "Select which mobs should trigger the enchantment (All by default)",
+    animal: "Select which animals should trigger the enchantment (All by default)",
+    items: "Select which items should trigger the enchantment (All by default)",
+    armor: "Select which armor should trigger the enchantment (All by default)",
+    prime_cause: "Select what tnt prime cause should trigger the enchantment (All by default)",
+    damage_causes: "Select which damage cause should trigger the enchantment (All by default)",
+    empty: "Fill in what exact values (player name, number...) should trigger the enchantment (All by default)",
+  };
