+++ frontend/src/data/versions.js
@@ -1,0 +1,1 @@
+export const versions = ["1.21.3", "1.21.1"]
