+++ frontend/src/styles/custom_enchants/forms.css
@@ -31,7 +31,7 @@
 }
 
 .input-field-text {
-  width: 200px
+  width: 70%
 }
 
 .input-field-number {
